requirements.txt
streamlit
llama-cpp-python
pyyaml